requirements/package.txt
django
django-countries
django-celery
multiprocessing
python-dateutil
pygeoip
pycurl
lxml
grab==0.4.8
